Fix the mapping of components to resources/category in the library
Each component is a row within a csv file, called a resource from the
datapackage perspective. For adding components automatically it is
important to know in which resource a component is located, thus this
mapping provided by the function `utils.list_available_components()`.
The module can be run standalone `python utils.py` to detect potential
problems each time one introduces a new component into the library.